pyproject.toml
[tool.poetry]
name = "revisao-rhnr"
version = "0.1.0"
description = ""
authors = ["mvcastro <mvcastro1975@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.43"
pandas = "^2.3.3"
python-dotenv = "^1.1.1"
psycopg2 = "^2.9.10"
xlrd = "^2.0.2"
openpyxl = "^3.1.5"
streamlit = "^1.50.0"


[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.3.2.250926"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
